ajout du tas

+++ src/math.cpp
@@ -1,6 +1,6 @@
 #include "math.hpp"
 
-uint64_t power_int(uint32_t x, uint32_t y) {
+int power_int(int x, int y) {
     int res = 1;
     while (y-- > 0)
         res *= x;
