+++ src/math.hpp
@@ -3,6 +3,6 @@
 
 #include "types.hpp"
 
-uint64_t power_int(uint32_t x, uint32_t y); //compute x^y
+int power_int(int x, int y); //compute x^y
 
 #endif
